Added changes to handle tier execution due to code refactoring

+++ cypress/integration/tests/applicationinventory/application/assess_review.test.ts
@@ -41,6 +41,8 @@ describe("Application assessment and review tests", { tags: "@tier1" }, () => {
     });
 
     after("Perform test data clean up", function () {
+        if (hasToBeSkipped("@tier1")) return;
+
         // Delete the stakeholders created before the tests
         deleteAllStakeholders();
     });
