+++ cypress/integration/tests/controls/businessservices/filter.test.ts
@@ -48,6 +48,9 @@ describe("Business services filter validations", { tags: "@tier2" }, function ()
     });
 
     after("Perform test data clean up", function () {
+        // Prevent hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier2")) return;
+
         // Delete the business services
         deleteAllBusinessServices();
         deleteAllStakeholders();
