+++ cypress/integration/tests/controls/businessservices/sort.test.ts
@@ -46,8 +46,10 @@ describe("Business services sort validations", { tags: "@tier2" }, function () {
     });
 
     after("Perform test data clean up", function () {
-        // Delete the bussiness services and stakeholders created before the tests
+        // Prevent hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier2")) return;
 
+        // Delete the bussiness services and stakeholders created before the tests
         deleteAllBusinessServices();
         deleteAllStakeholders();
     });
