+++ cypress/integration/tests/controls/stakeholdergroups/filter.test.ts
@@ -59,6 +59,9 @@ describe("Stakeholder groups filter validations", { tags: "@tier2" }, function (
     });
 
     after("Perform test data clean up", function () {
+        // Prevent before hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier2")) return;
+
         // Delete the stakeholder groups and stakeholders created before the tests
         deleteAllStakeholders();
         deleteAllStakeholderGroups();
