+++ cypress/integration/tests/controls/stakeholders/sort.test.ts
@@ -59,6 +59,9 @@ describe("Stakeholder sort validations", { tags: "@tier2" }, function () {
     });
 
     after("Perform test data clean up", function () {
+        // Prevent hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier2")) return;
+
         // Delete the job functions, stakeholder groups and stakeholders created before the tests
         deleteAllJobfunctions();
         deleteAllStakeholders();
